JENI-8: Pulled the private move functions out of the ViewModel and gave them their own file to make them testable. Wrote a test for the fixed bug of the random move path returning the piece's current position if it has no moves in JENI-5.
