+++ app/src/main/java/com/example/myapplication/GameViewModel.kt
@@ -107,12 +107,14 @@ class GameViewModel(
         }
 
         val newPositions = allyPositions.toMutableList()
+        val shuffledIndexes = (0 until allyPieces.size).toList().shuffled()
         val positionIndexPair = randomMove(
             turn = turn,
             enemyPositions = enemyPositions,
             enemyPieces = enemyPieces,
             allyPositions = allyPositions,
-            allyPieces = allyPieces
+            allyPieces = allyPieces,
+            shuffledAllyIndexes = shuffledIndexes
         )
         val newPosition = positionIndexPair.first
         // a stalemate happens when a player has no moves
@@ -153,82 +155,6 @@ class GameViewModel(
             Set.BLACK -> _gameState.value = _gameState.value.copy(
                 positionsBlack = mutableAllyPositions
             )
-        }
-    }
-
-    private fun randomMove(
-        turn: Set,
-        enemyPositions: List<List<Int>>,
-        enemyPieces: List<Piece>,
-        allyPositions: List<List<Int>>,
-        allyPieces: List<Piece>
-    ): Pair<List<Int>, Int> {
-        val pieceIndexes = (0 until allyPieces.size).toList().shuffled()
-        var newPosition: List<Int> = emptyList()
-        var newPositionIndex = 0
-        for (i in 0 until pieceIndexes.size) {
-            val position = randomNextPosition(
-                allyPieces[pieceIndexes[i]],
-                turn,
-                allyPositions[pieceIndexes[i]],
-                enemyPositions,
-                enemyPieces,
-                allyPositions
-            )
-            if (position.isNotEmpty()) {
-                newPosition = position
-                newPositionIndex = pieceIndexes[i]
-                break
-            } else if (i == pieceIndexes.size - 1) {
-                emptyList<Int>()
-            }
-        }
-
-        return Pair(newPosition, newPositionIndex)
-    }
-
-    private fun randomNextPosition(
-        piece: Piece,
-        turn: Set,
-        currentPosition: List<Int>,
-        enemyPositions: List<List<Int>>,
-        enemyPieces: List<Piece>,
-        allyPositions: List<List<Int>>
-    ): List<Int> {
-        val possibleMoves = piece.getValidMovesPositions(
-            Pair(currentPosition[0], currentPosition[1]), enemyPositions, allyPositions
-        )
-        if (possibleMoves.isEmpty()) return emptyList()
-
-        val teamPositions = allyPositions - currentPosition
-
-        val validMoves = possibleMoves.filter { move ->
-            val newPosition = listOf(move[0], move[1])
-            val validPosition = newPosition[0] in 0..7 &&
-                    newPosition[1] in 0..7 &&
-                    newPosition !in teamPositions
-            val kingDead = if (piece is King) {
-                piece.amIDead(
-                    position = Pair(move[0], move[1]),
-                    enemyPositions = enemyPositions,
-                    enemyPieces = enemyPieces,
-                    allyPositions = allyPositions,
-                )
-            } else { false }
-
-            validPosition && !kingDead
-        }
-
-        if (validMoves.isEmpty()) return emptyList()
-
-        // Prioritize capturing enemy King
-        val enemyKingIndex = enemyPieces.indexOfFirst { it is King }
-        val kingKillMove = validMoves.find { it == enemyPositions[enemyKingIndex] }
-        return if(enemyKingIndex != -1 && kingKillMove != null) {
-            println("${turn.name} ${piece.name} takes King at $kingKillMove!")
-            kingKillMove
-        } else {
-            validMoves.random()
         }
     }
